refactor: adjust syntax errors
