+++ src/pages/privileges/outlets/users/complete-invitation/interface.tsx
@@ -11,7 +11,7 @@ import { DecisionModal } from "@components/feedback/DecisionModal";
 import { ItemNotFound } from "@components/layout/ItemNotFound";
 import { PageTitle } from "@components/PageTitle";
 
-import { InitializerForm } from "../edit-user/forms/InitialiserForm";
+import { InitializerForm } from "../edit-user/forms/InitializerForm";
 import { GeneralInformationForm } from "../edit-user/forms/GeneralInfoForm";
 import {
   IAssignmentFormEntry,
